Seperate UTData extension methods into own class
